Removed and renamed server-config
