+++ server/app/src/router/fannyPack/addFannyPack/index.js
@@ -41,11 +41,11 @@ const TokenGenerator = require('uuid-token-generator');
 const Token = new TokenGenerator(); // New Token
 const moment = require('moment'); // Time
 // blingBlaw
-const create_schema_user_fannyPack = require("../../../modules/statements/fannyPack/addFannyPack/create_schema_user_fannyPack");
-const add_newFannyPack_to_fannypacks_table = require("../../../modules/statements/fannyPack/addFannyPack/add_newFannyPack_to_fannypacks_record");
-const create_table_account_types = require("../../../modules/statements/account/accountType/createAccountType/create_table_account_types");
-const create_table_account_category = require("../../../modules/statements/account/accountCategory/createAccountCategory/create_table_account_category");
-const create_table_account_records = require("../../../modules/statements/account/accountRecord/createAccountRecord/create_table_account_records");
+const create_schema_user_fannyPack = require("../../../config/statements/fannyPack/addFannyPack/create_schema_user_fannyPack");
+const add_newFannyPack_to_fannypacks_table = require("../../../config/statements/fannyPack/addFannyPack/add_newFannyPack_to_fannypacks_record");
+const create_table_account_types = require("../../../config/statements/account/accountType/createAccountType/create_table_account_types");
+const create_table_account_category = require("../../../config/statements/account/accountCategory/createAccountCategory/create_table_account_category");
+const create_table_account_records = require("../../../config/statements/account/accountRecord/createAccountRecord/create_table_account_records");
 // pageMessage
 let pageMessage = {
     title: "add_fannyPack",
